requirements.txt
google-generativeai
ipython
gtts
SpeechRecognition
pyaudio
